move connect storage function to utils
