+++ app/airflow/dags/libs/utils.py
@@ -2,14 +2,18 @@ import logging
 import ast
 import json
 from airflow.providers.postgres.hooks.postgres import PostgresHook
-from enum import Enum
 from airflow.operators.python import PythonOperator
 from typing import TypedDict, List, Optional
 from airflow.models.taskinstance import TaskInstance
-from libs.enums import JobStageType, StageStatusType
+from libs.enums import JobStageType, StageStatusType, StorageType
+import os
+from airflow.utils.session import create_session
+from airflow.models.connection import Connection
 
 # PostgreSQL connection hook
 pg_hook = PostgresHook(postgres_conn_id="postgres_db_conn")
+# Storage type
+storage_type = os.getenv("STORAGE_TYPE", StorageType.MINIO)
 
 
 # Define a type for field-vocab pairs
@@ -237,3 +241,55 @@ def pull_validated_params(kwargs: dict, task_id: str) -> ValidatedParams:
     """Pull parameters from XCom for a given task"""
     task_instance: TaskInstance = kwargs["ti"]
     return task_instance.xcom_pull(task_ids=task_id)
+
+
+def connect_to_storage() -> None:
+    """
+    Connects to the storage service based on the storage type.
+    """
+    with create_session() as session:
+        if storage_type == StorageType.AZURE:
+            # Check if WASB connection exists
+            existing_conn = (
+                session.query(Connection)
+                .filter(Connection.conn_id == "wasb_conn")
+                .first()
+            )
+
+            if existing_conn is None:
+                conn = Connection(
+                    conn_id="wasb_conn",
+                    conn_type="wasb",
+                    extra={
+                        "connection_string": os.getenv(
+                            "AIRFLOW_VAR_WASB_CONNECTION_STRING"
+                        ),
+                    },
+                )
+                session.add(conn)
+                session.commit()
+                logging.info("Created new WASB connection")
+
+        elif storage_type == StorageType.MINIO:
+            # Check if MinIO connection exists
+            existing_conn = (
+                session.query(Connection)
+                .filter(Connection.conn_id == "minio_conn")
+                .first()
+            )
+
+            if existing_conn is None:
+                conn = Connection(
+                    conn_id="minio_conn",
+                    conn_type="aws",
+                    extra={
+                        "endpoint_url": os.getenv("AIRFLOW_VAR_MINIO_ENDPOINT"),
+                        "aws_access_key_id": os.getenv("AIRFLOW_VAR_MINIO_ACCESS_KEY"),
+                        "aws_secret_access_key": os.getenv(
+                            "AIRFLOW_VAR_MINIO_SECRET_KEY"
+                        ),
+                    },
+                )
+                session.add(conn)
+                session.commit()
+                logging.info("Created new MinIO connection")
